refactor: move `ParentNamesView` into `birth` module
